requirements.txt
Pillow
anthropic
mss
python-dotenv